refactor: break duke-telemetry lib.rs blob into modules
